requirements.txt
anthropic
python-dotenv
tavily-python
Pillow
anthropic
rich
prompt_toolkit
pydub
websockets
SpeechRecognition